🤖 watchdogによるUSB監視部分を別モジュールに切り出し
- voice_monitor.pyモジュールを作成して監視ロジックを分離
- VoiceMemoAppクラスをシンプル化
- メインファイルのインポートを更新
- DEVELOPMENT.mdを更新
