+++ DEVELOPMENT.md
@@ -58,7 +58,9 @@ pip install rumps watchdog
 
 - [x] ボイスメモをコピー機能の実装 (voice_sync.pyの統合)
 - [x] テキスト起こし機能の実装 (voice_trans.pyの統合)
-- [x] USBデバイス自動検出機能の実装（watchdogライブラリを使用）
+- [x] USBデバイス自動検出機能の実装
+  - [x] watchdogライブラリを使用したファイルシステム監視
+  - [x] voice_monitor.pyモジュールとして分離して実装
 - [x] ボイスレコーダー接続時の自動処理機能の実装
 - [x] メニューから手動でUSBデバイス自動検出の有効/無効を切り替え可能に
 - [x] コマンドライン引数で起動時のUSBデバイス自動検出を有効化可能に
